Move components outside route- & UI-folder
