+++ src/components/TOS/Attribute.js
