+++ src/components/TOS/Attribute.scss
