+++ src/components/TOS/DeleteView.js
