+++ src/components/TOS/DeleteView.scss
