+++ src/components/TOS/HomeView.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import './HomeView.scss';
 
 export class HomeView extends React.Component {
   render () {
