+++ src/components/TOS/Record.js
