+++ src/layouts/CoreLayout/CoreLayout.js
@@ -15,7 +15,7 @@ export const CoreLayout = ({ children }) => (
 );
 
 CoreLayout.propTypes = {
-  children: React.PropTypes.element.isRequired
+  children: React.PropTypes.element
 };
 
 export default CoreLayout;
