--- src/layouts/CoreLayout/index.js
@@ -1,3 +1,0 @@
-import CoreLayout from './CoreLayout';
-
-export default CoreLayout;
