--- src/routes/UI/components/HomeView.scss
@@ -1,90 +1,0 @@
-@import '../../../styles/vendor/hel-components/variables';
-
-.navigation-menu {
-  padding: 15px;
-  background-color: white;
-  border: 1px solid $black50;
-  border-radius: $border-radius-base;
-}
-
-.navigation-menu-items {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  position: relative;
-}
-
-.nav-title {
-  font-size: 1.8rem;
-  letter-spacing: 0.03rem;
-  margin-bottom: 0.5rem;
-  text-decoration: underline;
-  margin: 0 0 0 35px;
-  line-height: 3.4rem;
-  display: block;
-  position: relative;
-}
-
-.nav-title:after {
-  content: '';
-  width: 9px;
-  height: 9px;
-  position: absolute;
-  left: -15px;
-  border-radius: 50%;
-  background: $black90;
-  top: 50%;
-  margin-top: -5px;
-}
-
-.nav-title:before {
-  left: -35px;
-  position: absolute;
-  content: '';
-  border-left: 1px solid $black90;
-  border-bottom: 1px solid $black90;
-  top: -50%;
-  width: 20px;
-  height: 100%;
-}
-
-.nav-main-title {
-  padding: 0 0 0 10px;
-  position: relative;
-}
-
-.nav-main-title:before {
-  content: '';
-  left: 0;
-  top: 20px;
-  border-left: 1px solid $black90;
-  position: absolute;
-  height: 20px;
-}
-
-.nav-main-title:after {
-  width: 9px;
-  height: 9px;
-  position: absolute;
-  left: -4px;
-  top: 14px;
-  content: '';
-  border-radius: 50%;
-  background: $black90;
-}
-
-.nav-section {
-  margin: -1.7rem 0 -1.7rem 24px;
-  padding: 1.7rem 0;
-  position: relative;
-}
-
-.nav-section:after {
-  position: absolute;
-  left: -24px;
-  border-left: 1px solid $black90;
-  height: 100%;
-  content: '';
-  top: 0;
-}
